not sure if the other commit went through

+++ src/main/java/com/example/bookclub/service/BookService.java
@@ -1,5 +1,6 @@
 package com.example.bookclub.service;
 
+import com.example.bookclub.exceptions.InformationNotFoundException;
 import com.example.bookclub.model.Book;
 import com.example.bookclub.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,9 +20,9 @@ public class BookService {
 //create book
     public Book createBook(Book bookObj){
         System.out.println("Service calling create book method ==> ");
-//        Book book = bookRepository.
         return bookRepository.save(bookObj);
     }
+    
 //get book list
     public List<Book> getBookList(){
         System.out.println("Service calling getBookList --> ");
@@ -33,28 +34,33 @@ public class BookService {
 
     public Optional<Book> getOneBook(Long bookId){
         System.out.println("Service calling getOneBook ==>");
-        return bookRepository.findById(bookId);
-        //        Optional<Book> book = bookRepository.findById(bookId);
-        //        if (book.isPresent()){
-        //            return book;
-        //        }else{
-        //            throw new InformationNotFoundException("book with id of " + researcherId +  " not found");
-        //        }
+        Optional<Book> book = bookRepository.findById(bookId);
+        if (book.isPresent()){
+            return book;
+        }else{
+            throw new InformationNotFoundException("book with id of " + bookId +  " not found");
+        }
     }
 
 //update one book
     public Book updateBook(Long bookId, Book bookObj){
         System.out.println("service calling updateBook ==> ");
-        Book updateBook = bookRepository.findById(bookId).get();
-        updateBook.setAuthor(bookObj.getAuthor());
-        updateBook.setGenre(bookObj.getGenre());
-        updateBook.setImg(bookObj.getImg());
-        updateBook.setLink(bookObj.getLink());
-        updateBook.setPages(bookObj.getPages());
-        updateBook.setTitle(bookObj.getTitle());
-        updateBook.setSummary(bookObj.getSummary());
-        updateBook.setOther(bookObj.getOther());
-        return bookRepository.save(updateBook);
+        Optional<Book> book = bookRepository.findById(bookId);
+        if(book.isPresent()) {
+            Book updateBook = bookRepository.findById(bookId).get();
+            updateBook.setAuthor(bookObj.getAuthor());
+            updateBook.setGenre(bookObj.getGenre());
+            updateBook.setImg(bookObj.getImg());
+            updateBook.setLink(bookObj.getLink());
+            updateBook.setPages(bookObj.getPages());
+            updateBook.setTitle(bookObj.getTitle());
+            updateBook.setSummary(bookObj.getSummary());
+            updateBook.setOther(bookObj.getOther());
+            return bookRepository.save(updateBook);
+        }
+        else {
+            throw new InformationNotFoundException("Book with id " + bookId + " not found");
+        }
         //        Optional<Book> updateBook = bookRepository.findById(bookId);
         //        if(updateBook.isPresent()) {
         //            updateBook.setAuthor(bookObj.getAuthor());
@@ -65,16 +71,20 @@ public class BookService {
 
 
 //delete one book
-    public Book deleteBook(Long bookId){
-        return bookRepository.deleteById(bookId);
-
+    public Optional<Book> deleteBook(Long bookId){
+        System.out.println("Service calling deleteBook ==> ");
+        Optional<Book> book = bookRepository.findById(bookId);
+        if(book.isPresent()) {
+            bookRepository.deleteById(bookId);
+            return book;
+        }else{
+            throw new InformationNotFoundException("Book with id " + bookId + " not found");
+        }
 //        Optional<Researcher> researcher = researcherRepository.findById(researcherId);
 //        if(researcher.isPresent()) {
 //            researcherRepository.deleteById(researcherId);
 //            return researcher;
-
         }
-}
